+++ src/ufo.rs
@@ -10,18 +10,18 @@ impl UFO {
         let width = 16 * PIXEL_SIZE as u32;
         let height = 7 * PIXEL_SIZE as u32;
 
-        let mut x = CANVAS_LEFT_EDGE;
+        let mut x = CANVAS_LEFT_EDGE as f32;
         let mut dir = String::from("right");
 
         if spawn_times % 2 != 0 {
-            x = CANVAS_RIGHT_EDGE;
+            x = CANVAS_RIGHT_EDGE as f32;
             dir = String::from("left")
         }
 
         UFO {
             game_object: GameObject::new(
                 x,
-                height as i32,
+                height as f32,
                 width,
                 height,
                 String::from("ufo_texture"),
@@ -32,9 +32,9 @@ impl UFO {
 
     pub fn move_x(&mut self) {
         if self.dir == "right" {
-            self.game_object.rect.x += 10;
+            self.game_object.rect.x += 10.0;
         } else {
-            self.game_object.rect.x -= 10;
+            self.game_object.rect.x -= 10.0;
         }
     }
 }
